Add training form page and 3 dashboard pages
I created a training form page and 3 dashboard pages (for admin, supervisor, and employee). They still have something to fix (e.g. make a CSS file to organize and make them neatly and consistently intended). Please check it when you have time. Also, I am still working on it to fix small bugs and to make it nicer. Again, they are not the final version, but pretty much all.